Fixup query context (#4979)
QueryContext::is_expired function never returns true if partial set of fragment instances are delivered, so the query is checked forever in pipeline poller threads. in addition, expired QueryContext cleanup depends on future QueryContexts cleanup, it is not a robust way, so add a QueryContext clean thread in QueryContextManager to scan expired QueryCoantexts and clean them.